% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scinference.R
\name{scinference}
\alias{scinference}
\title{Inference estimation_methods for synthetic control}
\usage{
scinference(
  Y1,
  Y0,
  T1,
  T0,
  inference_method = "conformal",
  alpha = 0.1,
  ci = FALSE,
  theta0 = 0,
  estimation_method = "sc",
  permutation_method = "mb",
  ci_grid = NULL,
  n_perm = 5000,
  lsei_type = 1,
  K = 2
)
}
\arguments{
\item{Y1}{outcome data for treated unit (T x 1 vector)}

\item{Y0}{outcome data for control units (T x J matrix)}

\item{T1}{number of post-treatment periods}

\item{T0}{number of pre-treatment periods}

\item{inference_method}{inference method; conformal inference ("conformal",default), t-test ("ttest")}

\item{alpha}{significance level; default \code{alpha} = 0.1}

\item{ci}{logical, indicating whether pointwise confidence intervals are reported; default \code{ci=FALSE}}

\item{theta0}{null hypothesis for treatment effect trajectory (T1 x 1 vector or scalar if constant effects null); default \code{theta0}=0}

\item{estimation_method}{estimation method; difference-in-differences ("did"), synthetic control ("sc", default), constrained lasso ("classo"). Note that constrained lasso is not implemented for the t-test.}

\item{permutation_method}{permutation method; moving block permutations ("mb", default), iid permutations ("iid")}

\item{ci_grid}{grid for the confidence interval}

\item{n_perm}{number of permutation (relevant for iid permutations); default = 5000}

\item{lsei_type}{option for lsei (package limSolve) used for sc; default = 1}

\item{K}{K>1 number of cross-fits for t-test; default = 2}
}
\value{
conformal inference: p-value for testing the null that \code{theta=theta0} and pointwise CI (lower bounds and upper bounds) if \code{ci=TRUE}; t-test: ATT estimate, standard error, and confidence interval
}
\description{
The function \code{scinference} implements the inference methods for synthetic controls proposed by Chernozhukov et al. (2020a,b).
The methods apply to a canonical synthetic control setup with 1 treated unit and J control units.
The treated unit is untreated for the first \code{T0} periods and treated for the remaining \code{T1=T-T0} periods.
}
